+++ sdk/src/main/java/com/eyeson/sdk/model/local/call/CameraOpen.kt
@@ -1,0 +1,5 @@
+package com.eyeson.sdk.model.local.call
+
+import com.eyeson.sdk.model.local.base.LocalBaseCommand
+
+internal data class CameraOpen(val cameraName: String?) : LocalBaseCommand